backend/requirements.txt
# FastAPI and related
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Environment
python-dotenv==1.0.0